[app] Initial commit of the starter app.
This commit of the starter app essentially aims to test how committing
a Java / Netbeans app works with Git. We can use NBGit to push changes
from within Netbeans. We can also load up the app/ folder directly as
a Netbeans project and it should ignore the generated local-specific
files and build files.