Worked on MainWindow, didn't change anything else. Tried to get remove working. Started out promising, but ran into some headaches and now its getting later and I'm about done looking at code for now. Can probably get it working pretty quick tomorrow though you'll probably have better ideas anyways.
I've got the removal of indexed lines working fully, though somewhere along the
way the compiler decided I was performing an unsafe action, not quite sure
what it was. It seems to run just fine, but every now and then it throws an
error while continuing to work, maybe its when I clear out _indexRecordList,
it was something about like 1 >= 1 or 6 >= 6 out of bounds exceptions or
something. :/

The bigger problems came with removing input data. I tried getting the index
of the selections so that I could then search through the _index and remove
all with origin_index equal to the selected indexes, but the indexes in the
_inputRecordList aren't the same as the indexes of the same strings in _index.
I'm not sure how we're going to go about this then other than just scanning
through all of the _index many times or changing things alot, maybe you've got
some ideas.

Anyways, I'll talk to you tomorrow, I'll get on gmail chat when I can and we
can hammer this thing out and finally put the nail in the coffin.