[search] Supports the  multiline search logic in the requirements.
The search has been modified to search so that every token on a single
line must be present for a match to occur. This is done for every line,
effectively combining the lines in an OR clause. This also brings in the
InputReader from the desktop application and updates it to be able to
take in StringReader and not perform validation (necessary for the new
environment).